Make it possible to build static juicysfplugin.so based on VST2.
Remove generated Projucer artifacts that do not build other than author's environment.

It is based on VST2 because JUCE does not support VST3 on Linux.

It avoids libfluidsynth.so because some? DAWs don't automatically loads
dependency shared libraries from VST directories. My Ubuntu desktop only has
Fluidsynth 1.x which is not what juicysfplugin expects.

Since it is not possible to build static fluidsynth with -fPIC yet, it is
based on my fork of fluidsynth.